requirements.txt
streamlit>=1.25.0
openai>=1.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
